Use exec for binary execution
Ensure that if run.sh is called directly with an existing mounted
/config, we can correctly execute the server.  This allows us to
manually set the uid/gid of steam and switch the entrypoint, causing the
server binary to be UID 1.  This allows graceful stopping of the server
when `docker stop` is issued.  Testing has unfortunately shown multiple
issues with propagating signals through the sudo call present in init.sh